The whole thing had a confliction
I don't know what happened, the whole thing had a confliction that wasn't able to be merged. So I had to fix them. Mostly deleted a ton of notes and extra stuff that didn't need to be there. I'm unsure, the code worked, like the site was doing great, it just was being weird.